backend/requirements.txt
# 基础依赖
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
torch
torchvision
torchaudio
faiss-cpu

# AI 和机器学习
huggingface-hub==0.16.4
transformers==4.36.2
sentence-transformers==2.2.2
google-generativeai

# Web 框架
fastapi
uvicorn
python-multipart
httpx

# 数据库
sqlalchemy
psycopg2-binary
alembic

# 文档处理
python-docx
PyPDF2
openpyxl
python-pptx

# 工具包
python-dotenv
tqdm
pandas

# 开发工具（可选）
black
flake8
pytest